requirements.txt
numpy
opencv-python
pillow
pytesseract
imutils
scikit-image
pathlib
facenet_pytorch
pandas
pyyaml
tqdm
matplotlib
seaborn
dlib
ffmpeg-python
wandb
flask
flask-cors
pywebview